[Update] AddMarkerUtil
