+++ app/src/main/java/com/bliends/pc/bliends/adapter/LocationAdapter.kt
@@ -6,8 +6,16 @@ import android.view.View
 import android.view.ViewGroup
 import com.bliends.pc.bliends.R
 import com.bliends.pc.bliends.data.Location
+import org.jetbrains.anko.find
 
-class LocationAdapter(private var mLocationLog: ArrayList<Location>) : RecyclerView.Adapter<LocationAdapter.ViewHolder>() {
+class LocationAdapter(private var mLocationLog: ArrayList<Location>) :
+        RecyclerView.Adapter<LocationAdapter.ViewHolder>(), View.OnClickListener {
+
+    override fun onClick(v: View?) {
+
+    }
+
+    private lateinit var view : View
 
     fun add(l: Location){
         mLocationLog.add(l)
@@ -19,6 +27,8 @@ class LocationAdapter(private var mLocationLog: ArrayList<Location>) : RecyclerV
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val v = LayoutInflater.from(parent.context).inflate(R.layout.location_log_item, parent, false)
+        view = v.rootView.find(R.id.location_log_item_layout)
+        view.setOnClickListener(this)
         return ViewHolder(v)
     }
