+++ app/src/main/res/layout/location_log_item.xml
@@ -3,7 +3,8 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:background="@drawable/help_log_line">
+    android:background="@drawable/help_log_line"
+    android:id="@+id/location_log_item_layout">
 
     <TextView
         android:id="@+id/message_text"
